fix(OBJ): .NET 4.6 compatibility (parsing)
Forced the OBJ loader to use the English CultureInfo instead of the one
defined by the operating system.

If .NET 4.6 (Experimental) is selected in Unity 2017.2 and above the
local culture info is used, causing the loader to parse float numbers
in a wrong way (when decimal separator is not the dot).