perf(ObjectBuilder): exploited the Unity 2017.3 support for large meshes
Added a platform dependent compilation to exploit the Unity 2017.3
support for large meshes.

If a mesh has more than 65K vertices/indices switch to 32 bit mode.

Note that in this version there is no check for the real support for 32
bit indices (see the comment at the beginning of `ObjectBuilder.cs`).