fix(ObjectBuilder/DataSet): check if no face is defined for an object
DataSet:
Discard the default object if no face was defined (even if vertices data
defined).

ObjectBuilder:
Added a check on the presence of faces in an object before creating its
meshes and renderers.
If no face is defined for an object an empty game object is created and
a warning is printed.

Added a check on the material library: in case of empty list create a
default material anyway.

Minor change in OBJ loader: added a TODO for implementing triangulation.