feat(Triangulation): triangulation of faces with ear-clipping algorithm
Starting from this
[tutorial](https://www.habrador.com/tutorials/math/10-triangulation/)
by Erik Nordeus, a triangulation of faces with more than 3 vertices has
been implemented, according to the "ear clipping" algorithm.

The previously mentioned code works only on the XZ plane, so a proper
projection was implemented to get vertices on the plane they lay on.

A refactoring and adaptation of the above code can be found in the
scripts in `MathUtil` folder, under the namespace `AsImpL.MathUtil`.

Updated documentation in `Documentation` folder.